- more on details view, but still not finished - introduced ajax controller to use common functionalities between pages (follow/unfollow, bookmark, etc) - the new reworked css is in use, but yet still looks ugly - created javascript views to make reusable JS calls over pages

+++ trunk/venefica-site/application/models/adstatistics_model.php
@@ -9,9 +9,9 @@ class AdStatistics_model extends CI_Model {
     
     var $numAvailProlongations = 0; //int
     var $numViews = 0; //long
-    var $numComments = 0; //long
-    var $numBookmarks = 0; //long
-    var $numShares = 0; //long
+    var $numComments = 0; //int
+    var $numBookmarks = 0; //int
+    var $numShares = 0; //int
     var $rating = 0; //float
     
     public function __construct($obj = null) {
